Ultimate Meshes Builder but with database info hardcoded in meshDatabase/Database/GameObjectHelper.cs Compile with x64 release and requires .NET 4.5
Selecting a gategory will build all tiles only for the selected category
To select the first option (BUILD ALL) which will build all tiles for all categories, scroll up
Selecting a zone name will build only the tiles of this zone

Currently configured to split each ADT in 4 tiles for next version 733
Includes all fixes (Doodad Quaternion and flags, WMO triangle groups filtering, Doodad transformations, gameobjects...)

+++ meshReader/RecastLayer/Recast/RecastAlloc.h
@@ -32,7 +32,7 @@ enum rcAllocHint
 //  @param[in]		rcAllocHint	A hint to the allocator on how long the memory is expected to be in use.
 //  @return A pointer to the beginning of the allocated memory block, or null if the allocation failed.
 ///  @see rcAllocSetCustom
-typedef void* (rcAllocFunc)(int size, rcAllocHint hint);
+typedef void* (rcAllocFunc)(size_t size, rcAllocHint hint);
 
 /// A memory deallocation function.
 ///  @param[in]		ptr		A pointer to a memory block previously allocated using #rcAllocFunc.
@@ -49,7 +49,7 @@ void rcAllocSetCustom(rcAllocFunc *allocFunc, rcFreeFunc *freeFunc);
 ///  @param[in]		hint	A hint to the allocator on how long the memory is expected to be in use.
 ///  @return A pointer to the beginning of the allocated memory block, or null if the allocation failed.
 /// @see rcFree
-void* rcAlloc(int size, rcAllocHint hint);
+void* rcAlloc(size_t size, rcAllocHint hint);
 
 /// Deallocates a memory block.
 ///  @param[in]		ptr		A pointer to a memory block previously allocated using #rcAlloc.
